Remove $wgCollectionHierarchyDelimiter workaround.
This was introduced in r73454 as a hacky workaround for lack of
subpage support in mwlib (see also T31648).  This lack has been
remedied in OCG (see T93008), so this workaround is no longer
required.

This is a breaking change for wikis using non-standard delimiters or
relying on pseudo-subpages without actually setting them as subpages,
but no such wiki is known (for instance Wikibooks uses the delimiter,
but also uses subpages).

This patch was forked from Federico Leva's work in
https://gerrit.wikimedia.org/r/163139, including some of his
improvements to code comments and type safety.

Change-Id: I924743f6afa45430ea6f77ffd96014aa63370a66