Concatenate pages and send to Electron
Fetches pages in a collection from RESTBase, sends them through
RemexHTML to fix heading levels and ensure id uniqueness, adds
cover page, table of contents and contributors, then outputs or
sends to Electron.

The functionality can be accessed at Special:RenderBook/test,
but (short of typing in the URL by hand) that is not
exposed to the user. Integrating it will happen in a
follow-up patch.

Bug: T171838
Change-Id: I686736e0b2c9f98a7c37046336219a0f852179b0
Depends-On: Ie37105a9f1ff92e1a79bfcd9f8578965e3d347f0