Refactor: Use mustache templates to create download box
The existing way of building HTML is prone to error and hard to
make sense of. In porting this to a Mustache template, I discovered
some invalid HTML relating to stray <tr><td> tags before a <tbody>

I've started with the download box, but as I touch other pieces of
code I'd like to also port them over.

Change-Id: Id5e3bfdc7ea5a7d908da2b9d904d97f794df844a