Hygiene: Render functions should contain no logic
Add the getBookTemplateData method which will be
responsible for creating the data needed for rendering.

Update method signatures so that they are only passed
what they need to render.

Add a comment to explain the additional processing in
renderImageInfos which is not clear at first glance.

Further simplification will happen in follow up patches

Change-Id: Iec31d60df0dd3d42402e781b5b1e356bc7e88800