Add strict, specific type hints were possible
This does not fix the entire codebase. I'm focusing on the special
page an a few closely related places. This mostly adds "array" type
hints to code that expects a $collection array, as well as "Title"
type hints. In one situation, no caller can ever pass null to the
method, making a sanity check obsolete.

Change-Id: I8845c58b8a368cf7a9d5a8cf37694e6c0d788d7c