Set defaults on arguments of CollectionAjaxFunctions class
The functions are called dynamically and the argument count depends on
the argument from a webrequest (ajax). When the webrequest does not
contains enough argument the server logs an error. It is better to act
with a default here and get errors from the underlying functions.
The argument counts could be less than expected when a user is trying to
call these functions with own code and not with the extension javascript

Bug: T197797
Change-Id: If9b719594f4040e24e5e5b13ddfe0926d06c379e