Fix special page layout
Book creation form makes use of clears but makes the assumption
the skin will clear for it. This is not a good assumption to make
as we don't know in what context the book creation form will be
rendered.

Bug: T280259
Change-Id: Ic273e3963921c106c23b72d4e6f90bf5f51274a0
(cherry picked from commit 72205bd0a36b12befa2274dd208b26bfc878e2bc)