Api: Avoid trying to access undefined offset in a user's collection
This was already there but wasn't showing up since the ajax interface
was only used by the extension UI and not in third party code like how
the new API modules are used now.

Maybe people had already forgotten how to use the AJAX interface (syntax)
and could no longer exploit this in code but with the new API modules
which is simpler to use, this is now showing up.

Try to rename a chapter that doesn't exist does nothing. Just return.

Bug: T293261
Change-Id: I2e1daba4cff53deac023089165bc79d5fdbce2a7