Drop support for RestbaseVirtualRestService
Collection tried to get page content for Special:RenderBook with parsiod
via RESTBase using RestbaseVirtualRestService and all these is dead code
as we can't even print PDFs of articles with this extension anymore.

So, Special:RenderBook got dropped entirely and any code related to it.
To add, since RestbaseVirtualRestService was used for this, it is dropped
as well.

Bug: T336735
Change-Id: I87834534434947e5cc251e5a4be5bea9e394f7f5