Fix some confusion with spaces in titles
The code handles titles containing spaces very inconsistently. Sometimes
it uses getPrefixedTest(), sometimes getDBKey(), leading to mismatches
in various places. The proper way to fix this would be to have callers
pass Title objects (or PageIdentity), and let whatever class sets the
data turn that into a string (IOW: encapsulation of the toString logic).

This patch fixes potential mismatches in Session::findArticle and a
couple failing tests, but the issue still exists.

Bug: T344461
Change-Id: I1b2073796d59850d6e6934ed33a8db39a7475606