Add basic guards to API calls that sometimes fail
I believe what happens here is that the API call fails (for whatever
reason), but there is no error handler in the code. It blindly
proceeeds with the happy path.

I believe it's reasonable to add trivial guards like this. The
alternative is to rewrite larger parts of this code. I don't think
this is worth it.

Bug: T345935
Change-Id: If5c08bb45723b174a5579db4d6c64c82d859d184