Fix Session class storing null in the list of 'items'
Issue most likely introduced by I63d9e19 and a design flaw in the
"setsorting" API:
a) It's an API call with input from a previous user session of
   unknown age. It's possible the data was modified in the meantime
   via another browser tab.
b) The API allows to delete entries. If the same API call is
   triggered again these entries are gone already, and the order of
   the rest might be totally different.

The very last issue I can't fix without an entirely new design (we
would need to pass titles to the API instead of numbers).

I try to rearrange the code and documentation a bit to hopefully
make it more obvious what's going on here.

Bug: T343470
Change-Id: I2f559d76da277c6e8a85410db0e10a3977904af7