Add missing safe guard to show/show_soon_timeout
Note how this code does a POST but has no error handler. It looks like
it always assumes the happy path and runs into an error if something
went wrong on the network.

Let's just bail out. It doesn't look like this is super critical
(calls a "getpopupdata" API module).

Bug: T266167
Change-Id: I1eaa96be52a47628bf55171b62a15db86e0eb6a0