Special:Book: Hide download-related options when download disabled
Make $wgCollectionDisableDownloadSection actually omit the
download section (instead of just disabling the form submit),
and also hide the form fields related to rendering settings
which are only relevant for downloading. (They are kept as
hidden fields because the options are passed around in many
different places and I didn't want to spend time on tracking
down whether something somewhere assumes they are set.)

This is very hacky but the extension is unmaintained so doing it
in a more proper way wouldn't be a good use of time.

Bug: T175681
Change-Id: Idea283762444ddf7a9e880c25cfc2b65894c2eb1